Move get_base_queryset / order_queryset to generic BaseListingView. Also handle the 'no ordering specified' case in the base order_queryset implementation, for consistency with filter_queryset and search_queryset.
